[fixup] prevent import cycle error errors in treestore
